+++ lib/src/m_login.dart
@@ -28,6 +28,10 @@ class MLogin {
   ///
   /// This URL scheme __MUST__ also be defined for Android in the
   /// `AndroidManifest.xml` file. See the README on how to do that.
+  ///
+  /// Some methods also have a parameter [overrideRedirectUri], which overrides
+  /// this general redirect uri for the given call.
+  ///
   final String redirectUri;
 
   /// Necessary to catch when the web process finishes in iOS.
@@ -40,16 +44,6 @@ class MLogin {
   /// Identifies the app using this MLoginSDK. Ask the MLogin team for a valid clientId.
   /// __MUST__ match the [redirectUri], access attempts will fail otherwise.
   final String clientId;
-
-  /// This is only used and required for the [openDriverLicenseVerification] method.
-  /// Defines where the (external) driver license verification service (e.g.,
-  /// IDNow) should redirect to after a successful verification step. You can
-  /// use the same redirect uri as set in [redirectUri] but it is recommended
-  /// to use a uri specific to this use case instead.
-  ///
-  /// Expected to be a full Uri, starting with your custom scheme.
-  /// Example: `my.custom.scheme:/redirect/id_verify`
-  final String? idVerificationRedirectUri;
 
   /// [loggedInMLoginUserId] Identifies the user that is currently logged-in in
   /// the app using the SDK. Optional, but it is __strongly__ recommended to set
@@ -72,6 +66,8 @@ class MLogin {
   /// __NOTE__: While it is allowed to have multiple instances, there should
   /// always be at most one call (thus: browser session) running at the same
   /// time.
+  /// The browser session is active for 15 minutes, or for 180 days, in case
+  /// the user set the "remain logged in" checkbox.
   ///
   /// [config] Defines which MLogin tier is to be accessed (i2, k, p)
   ///
@@ -104,20 +100,12 @@ class MLogin {
   ///
   /// Can (and should) be updated when the login changes.
   ///
-  /// [idVerificationRedirectUri] is only used and required for the
-  /// [openDriverLicenseVerification] method.
-  /// Defines where the (external) driver license verification service (e.g.,
-  /// IDNow) should redirect to after a successful verification step. You can
-  /// use the same redirect uri as set in [redirectUri] but it is recommended
-  /// to use a uri specific to this use case instead.
-  ///
   MLogin({
     required this.config,
     required this.redirectUri,
     required this.callbackUrlScheme,
     required this.clientId,
     this.loggedInMLoginUserId,
-    this.idVerificationRedirectUri,
     this.prefilledUsername,
   });
 
@@ -193,12 +181,16 @@ class MLogin {
   /// or validated. It's safe to ignore the returned value and just assume data
   /// was changed.
   ///
-  Future<bool> openPortalOverview({bool ephemeral = false}) {
+  Future<bool> openPortalOverview({
+    bool ephemeral = false,
+    String? overrideRedirectUri,
+  }) {
     return openDataPage(
       this,
       portalUriSuffix: 'profile',
       ephemeral: ephemeral,
       username: prefilledUsername,
+      overrideRedirectUri: overrideRedirectUri,
     );
   }
 
@@ -209,34 +201,75 @@ class MLogin {
   /// Does not require previous Login. In case there is no valid login present
   /// in the browser, the user will be prompted to log in again.
   ///
-  /// Requires [idVerificationRedirectUri] to be set!
-  /// Throws exception if not.
-  ///
   /// It is strongly recommended to keep [loggedInMLoginUserId] in sync with the
   /// logged in user of the calling app to ensure that the profile page is shown
   /// for the correct user.
+  ///
+  /// [idVerificationRedirectUri] defines where the (external)
+  /// driver license verification service (e.g. IDNow) should redirect to
+  /// after a successful verification step.
   ///
   /// Returns [true] in case the user finishes the page using the `done` button,
   /// [false] in any other case (e.g., the user pressed the "cancel" button in
   /// the iOS browser, or the back button on Android). This does not infer any
   /// data change or validation and can safely be ignored.
   ///
-  Future<bool> openDriverLicenseVerification({bool ephemeral = false}) {
-    final idRedirectUri = idVerificationRedirectUri;
-    if (idRedirectUri == null) {
-      throw Exception('"idVerificationRedirectUri" MUST be set when calling '
-          'openDriverLicenseVerification!');
-    }
-
+  Future<bool> openDriverLicenseVerification({
+    bool ephemeral = false,
+    required String idVerificationRedirectUri,
+  }) {
     return openDataPage(
       this,
       portalUriSuffix: 'driver-licence-direct-verification',
       extraParams: {
         'id_verification_client_id': clientId,
-        'id_verification_redirect_uri': idRedirectUri,
+        'id_verification_redirect_uri': idVerificationRedirectUri,
       },
       ephemeral: ephemeral,
       username: prefilledUsername,
+    );
+  }
+
+  ///
+  /// Opens the external photocollect library, which enables the user to
+  /// take a photo using her device or upload an existing picture.
+  /// The picture should be a portrait of the user, and is validated by the
+  /// library for correct format.
+  ///
+  /// Returns [true] if the portrait was successfully uploaded, and [false] if
+  /// anything went wrong or the user canceled the process.
+  ///
+  Future<bool> openPhotoUpload({
+    bool ephemeral = false,
+    String? overrideRedirectUri,
+  }) {
+    return openDataPage(
+      this,
+      portalUriSuffix: 'portrait',
+      ephemeral: ephemeral,
+      username: prefilledUsername,
+      overrideRedirectUri: overrideRedirectUri,
+    );
+  }
+
+  /// This function is used to authorize users for offers which are only
+  /// eligible for students, e.g. discounted tickets.
+  /// Shows a list of participating universities and colleges, which are
+  /// internally using the Shibboleth authentication.
+  /// User selects a university or college and is forwarded to its login.
+  ///
+  /// After successfully logging in, returns [true] and the user is
+  /// validated as a student. Returns [false] otherwise.
+  Future<bool> openStudentStatus({
+    bool ephemeral = false,
+    String? overrideRedirectUri,
+  }) {
+    return openDataPage(
+      this,
+      portalUriSuffix: 'student-status',
+      ephemeral: ephemeral,
+      username: prefilledUsername,
+      overrideRedirectUri: overrideRedirectUri,
     );
   }
 
@@ -246,8 +279,8 @@ class MLogin {
 
   ///
   /// Opens a page in the browser where the user can grant a SEPA mandate for
-  /// payments to the [payeeId] of the service, using the already entered bank
-  /// account information in the M-Login.
+  /// payments, using the already entered bank account information in the
+  /// M-Login.
   ///
   /// Calling this is usually the result of trying to trigger a payment for the
   /// given service ("checkout") and receiving a `mandate_missing` error message
@@ -261,8 +294,6 @@ class MLogin {
   /// Parameters, and where to get them:
   /// * The [methodId] identifies the SEPA payment method that needs a mandate.
   ///   If you tried to do a simple checkout, this is the default payment method
-  /// * The [payeeId] identifies the entity that shall receive the money. Talk
-  ///   to the M-Login team if you don't know what to put there
   ///
   /// Returns [true] in case the user finishes the page using the `done` button,
   /// [false] in any other case (e.g., the user pressed the "cancel" button in
@@ -272,14 +303,18 @@ class MLogin {
   /// or validated. It's safe to ignore the returned value and just assume data
   /// was changed.
   ///
-  Future<bool> openGrantSepaMandatePage(String methodId, String payeeId,
-      {bool ephemeral = false}) {
+  Future<bool> openGrantSepaMandatePage(
+    String methodId, {
+    bool ephemeral = false,
+    String? overrideRedirectUri,
+  }) {
     return openDataPage(
       this,
       portalUriSuffix: 'grantmandate',
-      extraParams: {'method_id': methodId, 'payee_id': payeeId},
+      extraParams: {'method_id': methodId},
       ephemeral: ephemeral,
       username: prefilledUsername,
+      overrideRedirectUri: overrideRedirectUri,
     );
   }
 
@@ -289,12 +324,6 @@ class MLogin {
   /// Does not require previous Login. In case there is no valid login present
   /// in the browser, the user will be prompted to log in again.
   ///
-  /// The parameter [payeeId] is the id of the entity that will receive money,
-  /// when the user pays in this service using SEPA. Must be added here to
-  /// ensure that, when the user adds a new SEPA payment method, the correct
-  /// mandate is already requested as well (see documentation of
-  /// [openGrantSepaMandatePage]).
-  ///
   /// Returns [true] in case the user finishes the page using the `done` button,
   /// [false] in any other case (e.g., the user pressed the "cancel" button in
   /// the iOS browser, or the back button on Android)
@@ -303,14 +332,16 @@ class MLogin {
   /// or validated. It's safe to ignore the returned value and just assume data
   /// was changed.
   ///
-  Future<bool> openPaymentMethodsOverviewPage(String payeeId,
-      {bool ephemeral = false}) {
+  Future<bool> openPaymentMethodsOverviewPage({
+    bool ephemeral = false,
+    String? overrideRedirectUri,
+  }) {
     return openDataPage(
       this,
       portalUriSuffix: 'paymentmethods',
-      extraParams: {'payee_id': payeeId},
       ephemeral: ephemeral,
       username: prefilledUsername,
+      overrideRedirectUri: overrideRedirectUri,
     );
   }
 
@@ -346,8 +377,11 @@ class MLogin {
   /// [false] in any other case (e.g., the user pressed the "cancel" button in
   /// the iOS browser, or the back button on Android)
   ///
-  Future<bool> openPayAuthorizationErrorRecovery(String recoverableErrorPayload,
-      {bool ephemeral = false}) {
+  Future<bool> openPayAuthorizationErrorRecovery(
+    String recoverableErrorPayload, {
+    bool ephemeral = false,
+    String? overrideRedirectUri,
+  }) {
     // We need to translate the received error JSON to query parameters in order
     // to hand it over to the Portal.
     // However, we just include 'error' and all fields in 'details' to avoid running
@@ -381,6 +415,7 @@ class MLogin {
       extraParams: extraParams,
       ephemeral: ephemeral,
       username: prefilledUsername,
+      overrideRedirectUri: overrideRedirectUri,
     );
   }